Methods for Evaluating the CRPS of Various Distributions

Methods to the crps generic function from the scoringRules package for evaluating the (continuous) ranked probability score (CRPS) of many distribution objects from the distributions3 package.

# S3 method for Beta
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Bernoulli
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Binomial
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Erlang
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Exponential
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Gamma
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for GEV
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Geometric
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Gumbel
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for HyperGeometric
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Logistic
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for LogNormal
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for NegativeBinomial
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Normal
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Poisson
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for StudentsT
crps(y, x, drop = TRUE, elementwise = NULL, ...)

# S3 method for Uniform
crps(y, x, drop = TRUE, elementwise = NULL, ...)

Arguments

y

A distribution object, e.g., as created by Normal or Binomial.

x

A vector of elements whose CRPS should be determined given the distribution y.

drop

logical. Should the result be simplified to a vector if possible?

elementwise

logical. Should each distribution in y be evaluated at all elements of x (elementwise = FALSE, yielding a matrix)? Or, if y and x have the same length, should the evaluation be done element by element (elementwise = TRUE, yielding a vector)? The default of NULL means that elementwise = TRUE is used if the lengths match and otherwise elementwise = FALSE is used.

...

Currently not used.

Value

In case of a single distribution object, either a numeric vector of length(x) (if drop = TRUE, default) or a matrix with length(x) columns (if drop = FALSE). In case of a vectorized distribution object, a matrix with length(x) columns containing all possible combinations.

Details

Methods for most univariate distributions are provided for which scoringRules provides a CRPS function and distributions3 provides a distribution object. For example, the crps method for Normal objects simply leverages crps_norm for objects of class Normal.

Examples

 if(!requireNamespace("distributions3")) {
  if(interactive() || is.na(Sys.getenv("_R_CHECK_PACKAGE_NAME_", NA))) {
    stop("not all packages required for the example are installed")
  } else q() }

set.seed(6020)

## three normal distributions X and observations x
library("distributions3")
#> 
#> Attaching package: ‘distributions3’
#> The following object is masked from ‘package:stats’:
#> 
#>     Gamma
#> The following object is masked from ‘package:grDevices’:
#> 
#>     pdf
X <- Normal(mu = c(0, 1, 2), sigma = c(2, 1, 1))
x <- c(0, 0, 1)

## evaluate crps using infrastructure from scoringRules
library("scoringRules")
crps(X, x)
#> [1] 0.4673900 0.6024414 0.6024414

## analogously for Poisson distribution
Y <- Poisson(c(0.5, 1, 2))
crps(Y, x)
#> [1] 0.1631650 0.4762224 0.4991650