(Randomized) Quantile Residuals

Generic function and methods for computing (randomized) quantile residuals.

qresiduals(object, ...)

# S3 method for default
qresiduals(
  object,
  newdata = NULL,
  trafo = qnorm,
  type = c("random", "quantile"),
  nsim = 1L,
  delta = NULL,
  prob = NULL,
  ...
)

Arguments

object
an object. For the default method this needs to be either a specification of probabilities (vector or 2-dimensional matrix of probabilities) or an object from which the these can be obtained with procast.

...
further parameters passed to methods.

newdata
optionally, a data frame in which to look for variables with which to predict. If omitted, the original observations are used.

trafo
function for tranforming residuals from probability scale to a different distribution scale (default: Gaussian).

type
character specifying whether - in the case of discrete response distributions - randomized quantile residuals or their corresponding quantiles should be computed.

nsim
numeric. The number of simulated randomized quantile residuals per observation (for type = "numeric").

delta
numeric. The minimal difference to compute the range of proabilities corresponding to each observation according to get (randomized) quantile residuals. For NULL, the minimal observed difference in the resonse divided by 5e-6 is used.

prob
numeric. The probabilities at which quantile residuals should be computed (for type = "quantile"), defaulting to the median (0.5).

Value

A vector or matrix of quantile residuals.

Details

(Randomized) quantile residuals have been suggested by Dunn and Smyth (1996). For regression models with a continuous response distribution this simply computes theoretical standard normal quantiles corresponding to the probability integral transform of the fitted distribution. For discrete distributions, a random theoretical normal quantile is drawn from the range of probabilities corresponding to each observation. Hence, in qqrplot the default is to use trafo = qnorm but other transformations can also be used, specifically using the uniform probability scale (via trafo = NULL or equivalently qunif or identity).

The default qresiduals method can compute randomized quantile residuals from a vector (which essentially just calls qnorm) or a 2-column matrix of probabilities. The latter offers to either draw "random" samples from the distribution or compute corresponding "quantile"s such as the median etc.

Note

Note that there is also a qresiduals function in the statmod package that is not generic and always returns a single random quantile residual.

References

Dunn KP, Smyth GK (1996). “Randomized Quantile Residuals.” Journal of Computational and Graphical Statistics, 5, 1--10.

See also

Examples

## linear regression models (homoscedastic Gaussian response)
m <- lm(dist ~ speed, data = cars)
qresiduals(m)

#>           1           2           3           4           5           6 
#>  0.25545799  0.78635430 -0.39470590  0.79981078  0.14067589 -0.51845900 
#>           7           8           9          10          11          12 
#> -0.24852536  0.28237095  0.81326726 -0.57585006  0.15413237 -1.03589883 
#>          13          14          15          16          17          18 
#> -0.63772660 -0.37227844 -0.10683030 -0.50051703  0.03037927  0.03037928 
#>          19          20          21          22          23          24 
#>  0.82672373 -0.76147969 -0.09785931  1.49482961  2.82207038 -1.42061458 
#>          25          26          27          28          29          30 
#> -1.02244236  0.83569472 -0.88523279 -0.35433648 -1.14619545 -0.61529913 
#>          31          32          33          34          35          36 
#>  0.04832125 -0.74353772  0.18553082  1.51277158  2.04366788 -1.40267262 
#>          37          38          39          40          41          42 
#> -0.73905223  0.72091262 -1.92908344 -0.86729081 -0.60184267 -0.33639451 
#>          43          44          45          46          47          48 
#>  0.19450180 -0.19469945 -1.25200656 -0.45117661  1.00878824  1.07515027 
#>          49          50 
#>  2.86692531  0.28329129