Extract Observed Responses from New Data

Generic function and methods for computing (randomized) quantile residuals.

newresponse(object, ...)

# S3 method for default
newresponse(object, newdata, na.action = na.pass, ...)

# S3 method for glm
newresponse(object, newdata, na.action = na.pass, ...)

# S3 method for gamlss
newresponse(object, newdata = newdata, ...)

Arguments

object

a fitted model object. For the default method this needs to needs to be formula-based so that model.frame can be used to extract the response from the original data the model was fitted to or terms can be used to set up the response on newdata.

...

further parameters passed to methods.

newdata

optionally, a data frame in which to look for variables with which to predict. If omitted, the original observations are used.

na.action

function determining what should be done with missing values in newdata. The default is to employ NA.

Value

A vector of new responses.

Details

This is a convenience function to accompany procast, e.g., for computing predicted probabilities at the observed responses (also known as the probability integral transform, PIT).

See also

Examples

## linear regression models (homoscedastic Gaussian response)
m <- lm(dist ~ speed, data = cars)
newresponse(m)
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   2  10   4  22  16  10  18  26  34  17  28  14  20  24  28  26  34  34  46  26 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>  36  60  80  20  26  54  32  40  32  40  50  42  56  76  84  36  46  68  32  48 
#>  41  42  43  44  45  46  47  48  49  50 
#>  52  56  64  66  54  70  92  93 120  85 
#> attr(,"response_type")
#> [1] "continuous"
#> attr(,"weights")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1
newresponse(m, newdata = cars[1:3, ])
#>  1  2  3 
#>  2 10  4 
#> attr(,"response_type")
#> [1] "continuous"
#> attr(,"weights")
#> [1] 1 1 1