Procast: Probabilistic Forecasting

Generic function and methods for computing various kinds of probabilistic forecasts from (regression) models.

procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = "quantile",
  at = 0.5,
  drop = FALSE,
  ...
)

# S3 method for default
procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = c("quantile", "mean", "variance", "probability", "density", "loglikelihood",
    "distribution", "parameters", "kurtosis", "skewness"),
  at = 0.5,
  drop = FALSE,
  ...
)

procast_setup(
  pars,
  FUN,
  at = NULL,
  drop = FALSE,
  type = "procast",
  elementwise = NULL,
  ...
)

# S3 method for disttree
procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = c("quantile", "location", "scale", "parameter", "density", "probability"),
  at = 0.5,
  drop = FALSE,
  use_distfamily = TRUE,
  ...
)

# S3 method for distforest
procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = c("quantile", "location", "scale", "parameter", "density", "probability"),
  at = 0.5,
  drop = FALSE,
  use_distfamily = TRUE,
  ...
)

Arguments

object

a fitted model object. For the default method this needs to have a prodist method.

newdata

optionally, a data frame in which to look for variables with which to predict. If omitted, the original observations are used.

na.action

function determining what should be done with missing values in newdata. The default is to employ NA.

type

character specifying the type of probabilistic forecast to compute. In procast_setup the type is only used for nice labels of the returned data frame.

at

specification of values at which the forecasts should be evaluated, typically a numeric vector but possibly also a matrix or data frame. Additionally, at can be the character string "function" or "list", see details below.

drop

logical. Should the result be simplified to a vector if possible (by dropping the dimension attribute)? If FALSE a matrix is always returned.

...

further parameters passed to methods.

pars

a data frame of predicted distribution parameters.

FUN

function to be used for forecasts. Either of type FUN(pars, ...) or FUN(at, pars, ...), see details below.

elementwise

logical. Should each element of distribution only be evaluated at the corresponding element of at (elementwise = TRUE) or at all elements in at (elementwise = FALSE). Elementwise evaluation is only possible if the number of observations is length of at are the same and in that case a vector of the same length is returned. Otherwise a matrix is returned. The default is to use elementwise = TRUE if possible, and otherwise elementwise = FALSE.

use_distfamily

For intern use only, will not be supported in the future.

Value

Either a data.frame of predictions (in case of multivariate forecasts, or if drop = FALSE, default) or a vector (in case of a univariate forecast and additionally drop = TRUE). Unless atis the character string "function" or "list" in which case a (list of) function(s) is returned.

Details

The function procast provides a unified framework for probabilistic forcasting (or procasting, for short) based on probabilistic (regression) models, also known as distributional regression approaches. Typical types of predictions include quantiles, probabilities, (conditional) expectations, variances, and (log-)densities. Internally, procast methods typically compute the predicted parameters for each observation and then compute the desired outcome for the distributions with the respective parameters.

Some quantities, e.g., the moments of the distribution (like mean or variance), can be computed directly from the predicted parameters of the distribution while others require an additional argument at which the distribution is evaluated (e.g., the probability of a quantile or an observation of the response.

The default procast method leverages the S3 classes and methods for probability distributions from the distributions3 package. It proceeds in two steps: First, prodist is used to obtain the predicted probability distribution object. Second, the extractor methods such as quantile, cdf, etc. are used to compute quantiles, probabilities, etc. from the distribution objects.

Therefore, to enable procast for a certain type of model object, the recommended approach is to implement a prodist method which can then be leveraged. However, if the distributions3 package does not support the necessary probability distribution, then it may also be necessary to implement a new distribution objects, see apply_dpqr.

Before adopting the distributions3 framework as the recommended workflow for procasting, the package had taken a different approach which is described in the following. Note, however, that this will be discontinued when we have converted all procasting methods to the new workflow.

Old workflow: The function procast_setup is a convenience wrapper that makes setting up procast methods easier for package developers. It takes a data frame of predicted parameters pars and a function FUN which is to be evaluated at the parameters. This can either have the interface FUN(pars, ...) when the desired quantity can be predicted directly from the predicted parameters -- or the interface FUN(at, pars, ...) if an additional argument at is needed. procast_setup takes care of suitable expanding at to the dimensions of pars.

Examples

## linear regression models (homoscedastic Gaussian response)
m <- lm(dist ~ speed, data = cars)

## medians on observed data
procast(m)
#>     quantile
#> 1  -1.849460
#> 2  -1.849460
#> 3   9.947766
#> 4   9.947766
#> 5  13.880175
#> 6  17.812584
#> 7  21.744993
#> 8  21.744993
#> 9  21.744993
#> 10 25.677401
#> 11 25.677401
#> 12 29.609810
#> 13 29.609810
#> 14 29.609810
#> 15 29.609810
#> 16 33.542219
#> 17 33.542219
#> 18 33.542219
#> 19 33.542219
#> 20 37.474628
#> 21 37.474628
#> 22 37.474628
#> 23 37.474628
#> 24 41.407036
#> 25 41.407036
#> 26 41.407036
#> 27 45.339445
#> 28 45.339445
#> 29 49.271854
#> 30 49.271854
#> 31 49.271854
#> 32 53.204263
#> 33 53.204263
#> 34 53.204263
#> 35 53.204263
#> 36 57.136672
#> 37 57.136672
#> 38 57.136672
#> 39 61.069080
#> 40 61.069080
#> 41 61.069080
#> 42 61.069080
#> 43 61.069080
#> 44 68.933898
#> 45 72.866307
#> 46 76.798715
#> 47 76.798715
#> 48 76.798715
#> 49 76.798715
#> 50 80.731124
procast(m, drop = TRUE)
#>         1         2         3         4         5         6         7         8 
#> -1.849460 -1.849460  9.947766  9.947766 13.880175 17.812584 21.744993 21.744993 
#>         9        10        11        12        13        14        15        16 
#> 21.744993 25.677401 25.677401 29.609810 29.609810 29.609810 29.609810 33.542219 
#>        17        18        19        20        21        22        23        24 
#> 33.542219 33.542219 33.542219 37.474628 37.474628 37.474628 37.474628 41.407036 
#>        25        26        27        28        29        30        31        32 
#> 41.407036 41.407036 45.339445 45.339445 49.271854 49.271854 49.271854 53.204263 
#>        33        34        35        36        37        38        39        40 
#> 53.204263 53.204263 53.204263 57.136672 57.136672 57.136672 61.069080 61.069080 
#>        41        42        43        44        45        46        47        48 
#> 61.069080 61.069080 61.069080 68.933898 72.866307 76.798715 76.798715 76.798715 
#>        49        50 
#> 76.798715 80.731124 

## probability integral transform (PIT) on observed data
procast(m, type = "probability", at = cars$dist)
#>    probability
#> 1   0.59882105
#> 2   0.77948846
#> 3   0.34947760
#> 4   0.78337764
#> 5   0.55481406
#> 6   0.30573228
#> 7   0.40380745
#> 8   0.60898167
#> 9   0.78722684
#> 10  0.28630374
#> 11  0.56001935
#> 12  0.15506020
#> 13  0.26603741
#> 14  0.35764644
#> 15  0.45831809
#> 16  0.31192384
#> 17  0.51187296
#> 18  0.51187296
#> 19  0.79103581
#> 20  0.22780511
#> 21  0.46180709
#> 22  0.92848877
#> 23  0.99715432
#> 24  0.08197447
#> 25  0.15822377
#> 26  0.79355266
#> 27  0.19287583
#> 28  0.36422902
#> 29  0.13071077
#> 30  0.27329840
#> 31  0.51888085
#> 32  0.23314913
#> 33  0.57212323
#> 34  0.93085744
#> 35  0.97737792
#> 36  0.08466918
#> 37  0.23449593
#> 38  0.76001388
#> 39  0.02937176
#> 40  0.19772705
#> 41  0.27770161
#> 42  0.37085189
#> 43  0.57556956
#> 44  0.42435458
#> 45  0.10996564
#> 46  0.32922254
#> 47  0.83852322
#> 48  0.85392752
#> 49  0.99751523
#> 50  0.60932787

## log-likelihood contributions
procast(m, type = "density", at = cars$dist, log = TRUE)
#>      density
#> 1  -3.683304
#> 2  -3.948789
#> 3  -3.726760
#> 4  -3.959034
#> 5  -3.661479
#> 6  -3.781003
#> 7  -3.681627
#> 8  -3.690252
#> 9  -3.969453
#> 10 -3.811149
#> 11 -3.663383
#> 12 -4.167061
#> 13 -3.847193
#> 14 -3.718503
#> 15 -3.657458
#> 16 -3.772228
#> 17 -3.652423
#> 18 -3.652423
#> 19 -3.980046
#> 20 -3.930308
#> 21 -3.656576
#> 22 -4.724547
#> 23 -7.474739
#> 24 -4.620690
#> 25 -4.153766
#> 26 -3.987205
#> 27 -4.028125
#> 28 -3.712246
#> 29 -4.282586
#> 30 -3.833704
#> 31 -3.653100
#> 32 -3.917347
#> 33 -3.668502
#> 34 -4.750449
#> 35 -5.656737
#> 36 -4.596375
#> 37 -3.914155
#> 38 -3.901443
#> 39 -5.438234
#> 40 -4.013032
#> 41 -3.825843
#> 42 -3.706297
#> 43 -3.670138
#> 44 -3.670175
#> 45 -4.404389
#> 46 -3.749689
#> 47 -4.140453
#> 48 -4.206835
#> 49 -7.597225
#> 50 -3.690502

## log-likelihood sum
sum(procast(m, type = "density", at = cars$dist, log = TRUE))
#> [1] -206.599
logLik(m)
#> 'log Lik.' -206.5784 (df=3)

## medians on new data
nd <- data.frame(speed = c(10, 15, 20))
procast(m, newdata = nd)
#>   quantile
#> 1 21.74499
#> 2 41.40704
#> 3 61.06908

## different quantile for each observation
procast(m, newdata = nd, at = c(0.25, 0.5, 0.75), elementwise = TRUE)
#>   quantile
#> 1 11.37162
#> 2 41.40704
#> 3 71.44245

## all combinations of quantiles and observations
procast(m, newdata = nd, at = c(0.25, 0.5, 0.75), elementwise = FALSE)
#>     q_0.25    q_0.5   q_0.75
#> 1 11.37162 21.74499 32.11837
#> 2 31.03366 41.40704 51.78041
#> 3 50.69571 61.06908 71.44245