Procast: Probabilistic Forecasting

Generic function and methods for computing various kinds of forecasts from probabilistic (regression) models (probabilistic 4-casting).

procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = "quantile",
  at = 0.5,
  ...
)

procast_setup(pars, FUN, at = NULL, drop = FALSE, type = "procast", ...)

# S3 method for lm
procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = c("quantile", "location", "scale", "parameter", "density", "probability",
    "score"),
  at = 0.5,
  drop = FALSE,
  ...
)

# S3 method for crch
procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = c("quantile", "location", "scale", "parameter", "density", "probability",
    "score"),
  at = 0.5,
  drop = FALSE,
  ...
)

# S3 method for disttree
procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = c("quantile", "location", "scale", "parameter", "density", "probability",
    "score"),
  at = 0.5,
  drop = FALSE,
  use_distfamily = TRUE,
  ...
)

# S3 method for glm
procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = c("quantile", "location", "scale", "parameter", "density", "probability",
    "score"),
  at = 0.5,
  drop = FALSE,
  ...
)

# S3 method for distforest
procast(
  object,
  newdata = NULL,
  na.action = na.pass,
  type = c("quantile", "location", "scale", "parameter", "density", "probability",
    "score"),
  at = 0.5,
  drop = FALSE,
  use_distfamily = TRUE,
  ...
)

Arguments

object
a fitted model object. For the default method this needs to be formula-based so that model.frame can be used to extract the response from the original data the model was fitted to or terms can be used to set up the response on newdata.

newdata
optionally, a data frame in which to look for variables with which to predict. If omitted, the original observations are used.

na.action
function determining what should be done with missing values in newdata. The default is to employ NA.

type
character specifying the type of probabilistic forecast to compute. In procast_setup the type is only used for nice labels of the returned data frame.

at
specification of values at which the forecasts should be evaluated, typically a numeric vector but possibly also a matrix or data frame. Additionally, at can be the character string "function" or "list", see details below.

...
further parameters passed to methods.

pars
a data frame of predicted distribution parameters.

FUN
function to be used for forecasts. Either of type FUN(pars, ...) or FUN(at, pars, ...), see details below.

drop
logical. Should forecasts be returned in a data frame (default) or (if possible) dropped to a vector, see details below.

use_distfamily
For intern use only, will not be supported in the future.

Value

Either a data.frame of predictions (in case of multivariate forecasts, or if drop = FALSE, default) or a vector (in case of a univariate forecast and additionally drop = TRUE). Unless at is the character string "function" or "list" in which case a (list of) function(s) is returned.

Details

The function procast provides a unified framework for probabilistic 4-casting based on probabilistic (regression) models, also known as distributional regression approaches. Typical types of predictions include quantiles, probabilities, (conditional) expectations, variances, (log-)densities, or scores. Internally, procast methods typically compute the predicted parameters for each observation and then transform these to the desired outcome. Some quantities (e.g., expectations or variances) can be computed directly from the predicted parameters of the distribution while others require an additional argument at which the distribution is evaluated (e.g., the probability of a quantile or an observation of the response. The argument at can also be the character "function" or "list" so that a single function or list of functions is set up that can be evaluated at different values later on.

The function procast_setup is a convenience wrapper that make setting up procast methods easier for package developers. It takes a data frame of predicted parameters pars and a function FUN which is to be evaluated at the parameters. This can either have the interface FUN(pars, ...) when the desired quantity can be predicted directly from the predicted parameters -- or the interface FUN(at, pars, ...) if an additional argument at is needed. procast_setup takes care of suitable expanding at to the dimensions of pars and optionally setting up a (list of) function(s) to be returned.

Examples

## linear regression models (homoscedastic Gaussian response)
m <- lm(dist ~ speed, data = cars)

## medians on observed data
procast(m)

#>     quantile
#> 1  -1.849460
#> 2  -1.849460
#> 3   9.947766
#> 4   9.947766
#> 5  13.880175
#> 6  17.812584
#> 7  21.744993
#> 8  21.744993
#> 9  21.744993
#> 10 25.677401
#> 11 25.677401
#> 12 29.609810
#> 13 29.609810
#> 14 29.609810
#> 15 29.609810
#> 16 33.542219
#> 17 33.542219
#> 18 33.542219
#> 19 33.542219
#> 20 37.474628
#> 21 37.474628
#> 22 37.474628
#> 23 37.474628
#> 24 41.407036
#> 25 41.407036
#> 26 41.407036
#> 27 45.339445
#> 28 45.339445
#> 29 49.271854
#> 30 49.271854
#> 31 49.271854
#> 32 53.204263
#> 33 53.204263
#> 34 53.204263
#> 35 53.204263
#> 36 57.136672
#> 37 57.136672
#> 38 57.136672
#> 39 61.069080
#> 40 61.069080
#> 41 61.069080
#> 42 61.069080
#> 43 61.069080
#> 44 68.933898
#> 45 72.866307
#> 46 76.798715
#> 47 76.798715
#> 48 76.798715
#> 49 76.798715
#> 50 80.731124

procast(m, drop = TRUE)

#>         1         2         3         4         5         6         7         8 
#> -1.849460 -1.849460  9.947766  9.947766 13.880175 17.812584 21.744993 21.744993 
#>         9        10        11        12        13        14        15        16 
#> 21.744993 25.677401 25.677401 29.609810 29.609810 29.609810 29.609810 33.542219 
#>        17        18        19        20        21        22        23        24 
#> 33.542219 33.542219 33.542219 37.474628 37.474628 37.474628 37.474628 41.407036 
#>        25        26        27        28        29        30        31        32 
#> 41.407036 41.407036 45.339445 45.339445 49.271854 49.271854 49.271854 53.204263 
#>        33        34        35        36        37        38        39        40 
#> 53.204263 53.204263 53.204263 57.136672 57.136672 57.136672 61.069080 61.069080 
#>        41        42        43        44        45        46        47        48 
#> 61.069080 61.069080 61.069080 68.933898 72.866307 76.798715 76.798715 76.798715 
#>        49        50 
#> 76.798715 80.731124 

## probability integral transform (PIT) on observed data
procast(m, type = "probability", at = cars$dist)

#>    probability
#> 1   0.60081531
#> 2   0.78417003
#> 3   0.34652997
#> 4   0.78808978
#> 5   0.55593701
#> 6   0.30206903
#> 7   0.40186398
#> 8   0.61117046
#> 9   0.79196758
#> 10  0.28235827
#> 11  0.56124732
#> 12  0.15012467
#> 13  0.26182584
#> 14  0.35484277
#> 15  0.45746181
#> 16  0.30835554
#> 17  0.51211772
#> 18  0.51211772
#> 19  0.79580317
#> 20  0.22318530
#> 21  0.46102201
#> 22  0.93252054
#> 23  0.99761426
#> 24  0.07771442
#> 25  0.15328579
#> 26  0.79833667
#> 27  0.18801555
#> 28  0.36154337
#> 29  0.12585715
#> 30  0.26917860
#> 31  0.51926989
#> 32  0.22857810
#> 33  0.57359365
#> 34  0.93483116
#> 35  0.97950682
#> 36  0.08035724
#> 37  0.22993764
#> 38  0.76451836
#> 39  0.02686025
#> 40  0.19289135
#> 41  0.27363944
#> 42  0.36828670
#> 43  0.57710851
#> 44  0.42281413
#> 45  0.10528374
#> 46  0.32593113
#> 47  0.84346190
#> 48  0.85884627
#> 49  0.99792760
#> 50  0.61152323

## log-likelihood contributions
procast(m, type = "density", at = cars$dist, log = TRUE)

#>      density
#> 1  -3.664198
#> 2  -3.940745
#> 3  -3.709465
#> 4  -3.951417
#> 5  -3.641463
#> 6  -3.765968
#> 7  -3.662451
#> 8  -3.671435
#> 9  -3.962270
#> 10 -3.797370
#> 11 -3.643447
#> 12 -4.168112
#> 13 -3.834916
#> 14 -3.700864
#> 15 -3.637275
#> 16 -3.756827
#> 17 -3.632030
#> 18 -3.632030
#> 19 -3.973305
#> 20 -3.921494
#> 21 -3.636357
#> 22 -4.748826
#> 23 -7.613609
#> 24 -4.640642
#> 25 -4.154263
#> 26 -3.980761
#> 27 -4.023387
#> 28 -3.694346
#> 29 -4.288451
#> 30 -3.820865
#> 31 -3.632736
#> 32 -3.907993
#> 33 -3.648779
#> 34 -4.775808
#> 35 -5.719858
#> 36 -4.615314
#> 37 -3.904668
#> 38 -3.891426
#> 39 -5.492250
#> 40 -4.007665
#> 41 -3.812676
#> 42 -3.688149
#> 43 -3.650484
#> 44 -3.650523
#> 45 -4.415329
#> 46 -3.733349
#> 47 -4.140395
#> 48 -4.209543
#> 49 -7.741199
#> 50 -3.671696

## log-likelihood sum
sum(procast(m, type = "density", at = cars$dist, log = TRUE))

#> [1] -206.5784

logLik(m)

#> 'log Lik.' -206.5784 (df=3)

## medians on new data
nd <- data.frame(speed = c(10, 15, 20))
procast(m, newdata = nd)

#>   quantile
#> 1 21.74499
#> 2 41.40704
#> 3 61.06908

## different quantile for each observation
procast(m, newdata = nd, at = c(0.25, 0.5, 0.75))

#>   quantile
#> 1 11.58120
#> 2 41.40704
#> 3 71.23287

## all combinations of quantiles and observations
procast(m, newdata = nd, at = rbind(c(0.25, 0.5, 0.75)))

#>     q_0.25    q_0.5   q_0.75
#> 1 11.58120 21.74499 31.90878
#> 2 31.24325 41.40704 51.57083
#> 3 50.90529 61.06908 71.23287

## function for computing quantiles (vectorized)
qnt1 <- procast(m, newdata = nd, at = "function")
## as before
qnt1(0.5)

#>        1        2        3 
#> 21.74499 41.40704 61.06908 

qnt1(c(0.25, 0.5, 0.75))

#>        1        2        3 
#> 11.58120 41.40704 71.23287 

qnt1(rbind(c(0.25, 0.5, 0.75)))

#>     q_0.25    q_0.5   q_0.75
#> 1 11.58120 21.74499 31.90878
#> 2 31.24325 41.40704 51.57083
#> 3 50.90529 61.06908 71.23287

## list of functions
qnt2 <- procast(m, newdata = nd, at = "list")
qnt2[[1]]

#> function (at, mu = 21.7449927007299, sigma = 15.0688559957914, 
#>     ...) 
#> FUN(at = at, pars = data.frame(mu = mu, sigma = sigma), ...)
#> <environment: 0x555c8c9cdd90>

qnt2[[1]](0.5)

#> [1] 21.74499

qnt2[[1]](c(0.25, 0.5, 0.75))

#> [1] 11.58120 21.74499 31.90878